check for empty table

+++ testdb.py
@@ -50,33 +50,34 @@ async def remove_robinhoodimmune():
     try:
         statement = "SELECT * from users;"
         cursor.execute(statement)
-        for(uid, date, time) in cursor:
-            datetimenow = datetime.now()
-            addTime = timedelta(minutes=5)
-            dateNowStr = datetimenow.strftime("%Y-%m-%d")
-            timeNowStr = datetimenow.strftime("%H:%M:%S")
-            mDate = (datetime.strptime(dateNowStr, "%Y-%m-%d")).date()
-            mTime = datetime.strptime(timeNowStr, "%H:%M:%S")
-            newTime = (mTime - addTime).time()
-            uDate = date
-            uTimestr = str(time)
-            uTime = datetime.strptime(uTimestr, "%H:%M:%S").time()
-            uUid = int(uid)
-            print({uid},{date})
-            if uDate <= mDate:
-                if uTime >= newTime:
-                    member = client.get_user(uid)
-                    for guild in client.guilds:
-                        for member in guild.members:
-                            for role in member.roles:
-                                if role.name == "Robinhood Immune":
-                                    role  = discord.utils.get(member.guild.roles, name="Robinhood Immune")
-                                    await member.remove_roles(role)
-                                    remove = "DELETE FROM `users` WHERE uid = (%s)"
-                                    data = (uid,)
-                                    cursor.execute(remove, data)
-                                    connection.commit()
-                                    print("role removed")
+        if cursor.fetchone():
+            for(uid, date, time) in cursor:
+                datetimenow = datetime.now()
+                addTime = timedelta(minutes=5)
+                dateNowStr = datetimenow.strftime("%Y-%m-%d")
+                timeNowStr = datetimenow.strftime("%H:%M:%S")
+                mDate = (datetime.strptime(dateNowStr, "%Y-%m-%d")).date()
+                mTime = datetime.strptime(timeNowStr, "%H:%M:%S")
+                newTime = (mTime - addTime).time()
+                uDate = date
+                uTimestr = str(time)
+                uTime = datetime.strptime(uTimestr, "%H:%M:%S").time()
+                uUid = int(uid)
+                print({uid},{date})
+                if uDate <= mDate:
+                    if uTime >= newTime:
+                        member = client.get_user(uid)
+                        for guild in client.guilds:
+                            for member in guild.members:
+                                for role in member.roles:
+                                    if role.name == "Robinhood Immune":
+                                        role  = discord.utils.get(member.guild.roles, name="Robinhood Immune")
+                                        await member.remove_roles(role)
+                                        remove = "DELETE FROM `users` WHERE uid = (%s)"
+                                        data = (uid,)
+                                        cursor.execute(remove, data)
+                                        connection.commit()
+                                        print("role removed")
 
     except database.Error as e:
         print(f" remove {e}")
